+++ lib/jxl/render_pipeline/test_render_pipeline_stages.h
@@ -28,13 +28,13 @@ class UpsampleXSlowStage : public RenderPipelineStage {
       const float* row = GetInputRow(input_rows, c, 0);
       float* row_out = GetOutputRow(output_rows, c, 0);
       for (int64_t x = -xextra; x < (int64_t)(xsize + xextra); x++) {
-        float xp = row[x + kRenderPipelineXOffset - 1];
-        float xc = row[x + kRenderPipelineXOffset];
-        float xn = row[x + kRenderPipelineXOffset + 1];
+        float xp = *(row + x - 1);
+        float xc = *(row + x);
+        float xn = *(row + x + 1);
         float xout0 = xp * 0.25f + xc * 0.75f;
         float xout1 = xc * 0.75f + xn * 0.25f;
-        row_out[kRenderPipelineXOffset + 2 * x + 0] = xout0;
-        row_out[kRenderPipelineXOffset + 2 * x + 1] = xout1;
+        *(row_out + 2 * x + 0) = xout0;
+        *(row_out + 2 * x + 1) = xout1;
       }
     }
   }
@@ -59,13 +59,13 @@ class UpsampleYSlowStage : public RenderPipelineStage {
       float* row_out0 = GetOutputRow(output_rows, c, 0);
       float* row_out1 = GetOutputRow(output_rows, c, 1);
       for (int64_t x = -xextra; x < (int64_t)(xsize + xextra); x++) {
-        float xp = rowp[x + kRenderPipelineXOffset];
-        float xc = rowc[x + kRenderPipelineXOffset];
-        float xn = rown[x + kRenderPipelineXOffset];
+        float xp = *(rowp + x);
+        float xc = *(rowc + x);
+        float xn = *(rown + x);
         float yout0 = xp * 0.25f + xc * 0.75f;
         float yout1 = xc * 0.75f + xn * 0.25f;
-        row_out0[kRenderPipelineXOffset + x] = yout0;
-        row_out1[kRenderPipelineXOffset + x] = yout1;
+        *(row_out0 + x) = yout0;
+        *(row_out1 + x) = yout1;
       }
     }
   }
@@ -84,8 +84,7 @@ class Check0FinalStage : public RenderPipelineStage {
                   float* JXL_RESTRICT temp) const final {
     for (size_t c = 0; c < input_rows.size(); c++) {
       for (size_t x = 0; x < xsize; x++) {
-        JXL_CHECK(fabsf(GetInputRow(input_rows, c,
-                                    0)[x + kRenderPipelineXOffset]) < 1e-8);
+        JXL_CHECK(fabsf(GetInputRow(input_rows, c, 0)[x]) < 1e-8);
       }
     }
   }